src/requirements.txt
Django (2.0.4)
django-cors-headers (2.2.0)
djangorestframework (3.8.1)
djangorestframework-jwt (1.11.0)
djutil (0.2)
pip (9.0.3)
psycopg2 (2.7.4)
PyJWT (1.6.1)
pytz (2018.3)
setuptools (39.0.1)
wheel (0.31.0)
